# swagger.yaml
openapi: 3.0.3
info:
  title: Nexus Project Management API
  description: |
    RESTful API for the Nexus Project Management System.
    
    ## Authentication
    This API uses JWT (JSON Web Token) for authentication. After registering or logging in, 
    you will receive a JWT token that must be included in the `Authorization` header for all 
    protected endpoints.
    
    ## Authorization
    The API implements role-based access control (RBAC) with two roles:
    - **ADMIN**: Full access to all resources
    - **USER**: Limited access to assigned resources only
    
    ## Features
    - User authentication and management
    - Project management with team assignments
    - Task management with status tracking
    - Role-based authorization
    - Pagination and search capabilities
  version: 1.0.0
  contact:
    name: API Support
    email: support@nexus-pm.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.nexus-pm.com
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from login or register endpoints.
        
        Example: `Authorization: Bearer <your_jwt_token>`

  # ============================================
  # REUSABLE SCHEMAS
  # ============================================
  schemas:
...
    UserRole:
      type: string
      enum:
        - ADMIN
        - USER
      description: User role for authorization
    TaskStatus:
      type: string
      enum:
        - TODO
        - IN_PROGRESS
        - DONE
      description: Current status of a task
    TaskPriority:
      type: string
      enum:
        - LOW
        - MEDIUM
        - HIGH
      description: Priority level of a task

    # ========== AUTH SCHEMAS ==========
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
          description: Unique user identifier
        name:
          type: string
          example: John Doe
          description: User's full name
        email:
          type: string
          format: email
          example: john.doe@example.com
          description: User's email address
        role:
          $ref: '#/components/schemas/UserRole'
        createdAt:
          type: string
          format: date-time
          example: 2024-01-15T10:30:00Z
          description: Account creation timestamp
        updatedAt:
          type: string
          format: date-time
          example: 2024-01-15T10:30:00Z
          description: Last update timestamp

    # ========== PROJECT SCHEMAS ==========
    CreateProjectRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 255
          example: Website Redesign
          description: Project name
        description:
          type: string
          example: Complete redesign of the company website
          description: Detailed project description
        userIds:
          type: array
          items:
            type: string
            format: uuid
          example: ["550e8400-e29b-41d4-a716-446655440000"]
          description: Array of user IDs to assign to the project
    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
          example: Website Redesign v2
          description: Updated project name
        description:
          type: string
          example: Complete redesign with new branding
          description: Updated project description
        userIds:
          type: array
          items:
            type: string
            format: uuid
          example: ["550e8400-e29b-41d4-a716-446655440000"]
          description: Updated array of assigned user IDs
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
          description: Unique project identifier
        name:
          type: string
          example: Website Redesign
          description: Project name
        description:
          type: string
          nullable: true
          example: Complete redesign of the company website
          description: Project description
        createdById:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
          description: ID of the user who created the project
        createdAt:
          type: string
          format: date-time
          example: 2024-01-15T10:30:00Z
          description: Project creation timestamp
        updatedAt:
          type: string
          format: date-time
          example: 2024-01-15T10:30:00Z
          description: Last update timestamp
        createdBy:
          $ref: '#/components/schemas/User'
        users:
          type: array
          items:
            $ref: '#/components/schemas/ProjectUser'
          description: Users assigned to this project
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/Task'
          description: Tasks belonging to this project
    ProjectUser:
      type: object
      properties:
        projectId:
          type: string
          format: uuid
          description: Project ID
        userId:
          type: string
          format: uuid
          description: User ID
        user:
          $ref: '#/components/schemas/User'
    AssignUsersRequest:
      type: object
      required:
        - userIds
      properties:
        userIds:
          type: array
          items:
            type: string
            format: uuid
          minItems: 1
          example: ["550e8400-e29b-41d4-a716-446655440000", "660e8400-e29b-41d4-a716-446655440001"]
          description: Array of user IDs to assign or remove

    # ========== TASK SCHEMAS ==========
...
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
          description: Unique task identifier
        title:
          type: string
          example: Implement user authentication
          description: Task title
        description:
          type: string
          nullable: true
          example: Implement JWT-based authentication system
          description: Task description
        status:
          $ref: '#/components/schemas/TaskStatus'
        priority:
          $ref: '#/components/schemas/TaskPriority'
        dueDate:
          type: string
          format: date-time
          nullable: true
          example: 2024-02-01T00:00:00Z
          description: Task deadline
        projectId:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
          description: Associated project ID
        assigneeId:
          type: string
          format: uuid
          example: 660e8400-e29b-41d4-a716-446655440001
          description: Assigned user ID
        createdAt:
          type: string
          format: date-time
          example: 2024-01-15T10:30:00Z
          description: Task creation timestamp
        updatedAt:
          type: string
          format: date-time
          example: 2024-01-15T10:30:00Z
          description: Last update timestamp
        project:
          $ref: '#/components/schemas/Project'
        assignee:
          $ref: '#/components/schemas/User'
        history:
          type: array
          items:
            $ref: '#/components/schemas/TaskHistory'
          description: Task change history
    TaskHistory:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique history record identifier
        taskId:
          type: string
          format: uuid
          description: Associated task ID
        updatedById:
          type: string
          format: uuid
          description: ID of user who made the change
        oldStatus:
          $ref: '#/components/schemas/TaskStatus'
        newStatus:
          $ref: '#/components/schemas/TaskStatus'
        oldPriority:
          $ref: '#/components/schemas/TaskPriority'
        newPriority:
          $ref: '#/components/schemas/TaskPriority'
        timestamp:
          type: string
          format: date-time
          example: 2024-01-15T10:30:00Z
          description: When the change occurred
        updatedBy:
          $ref: '#/components/schemas/User'

    # ========== USER MANAGEMENT SCHEMAS ==========
...
    Error:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
          description: HTTP status code
        message:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
          example: Validation failed
          description: Error message or array of messages
        error:
          type: string
          example: Bad Request
          description: Error type
    UnauthorizedError:
      type: object
      properties:
        statusCode:
          type: integer
          example: 401
        message:
          type: string
          example: Unauthorized
        error:
          type: string
          example: Unauthorized
    ForbiddenError:
      type: object
      properties:
        statusCode:
          type: integer
          example: 403
        message:
          type: string
          example: Forbidden resource
        error:
          type: string
          example: Forbidden
    NotFoundError:
      type: object
      properties:
        statusCode:
          type: integer
          example: 404
        message:
          type: string
          example: Resource not found
        error:
          type: string
          example: Not Found

# ============================================
# API ENDPOINTS
# ============================================
paths:
...
  /projects:
    post:
      tags:
        - Projects
      summary: Create a new project
      description: |
        Create a new project with optional user assignments.
        
        **Authentication Required:** Yes (JWT)
        
        **Authorization:** ADMIN only
        
        - Only admins can create projects
        - Creator is automatically recorded
        - Users can be assigned during creation
      operationId: createProject
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
            examples:
              basicProject:
                summary: Basic project
                value:
                  name: Website Redesign
                  description: Complete redesign of the company website
              projectWithUsers:
                summary: Project with assigned users
                value:
                  name: Mobile App Development
                  description: New iOS and Android mobile application
                  userIds:
                    - 550e8400-e29b-41d4-a716-446655440000
                    - 660e8400-e29b-41d4-a716-446655440001
      responses:
        '201':
          description: Project created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '403':
          description: Forbidden - User role not permitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenError'
              examples:
                notAdmin:
                  summary: User is not an admin
                  value:
                    statusCode: 403
                    message: Forbidden resource
                    error: Forbidden
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      tags:
        - Projects
      summary: Get all projects
      description: |
        Retrieve all projects based on user role.
        
        **Authentication Required:** Yes (JWT)
        
        **Authorization:**
        - **ADMIN**: Can view all projects in the system
        - **USER**: Can only view projects they are assigned to
      operationId: getAllProjects
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Projects retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
              examples:
                adminView:
                  summary: Admin view (all projects)
                  value:
                    - id: 550e8400-e29b-41d4-a716-446655440000
                      name: Website Redesign
                      description: Complete redesign of the company website
                      createdById: 770e8400-e29b-41d4-a716-446655440002
                      createdAt: "2024-01-15T10:30:00Z"
                      updatedAt: "2024-01-15T10:30:00Z"
                    - id: 660e8400-e29b-41d4-a716-446655440001
                      name: Mobile App
                      description: iOS and Android app
                      createdById: 770e8400-e29b-41d4-a716-446655440002
                      createdAt: "2024-01-16T09:00:00Z"
                      updatedAt: "2024-01-16T09:00:00Z"
                userView:
                  summary: User view (assigned projects only)
                  value:
                    - id: 550e8400-e29b-41d4-a716-446655440000
                      name: Website Redesign
                      description: Complete redesign of the company website
                      createdById: 770e8400-e29b-41d4-a716-446655440002
                      createdAt: "2024-01-15T10:30:00Z"
                      updatedAt: "2024-01-15T10:30:00Z"
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /projects/{id}:
    get:
      tags:
        - Projects
      summary: Get a project by ID
      description: |
        Retrieve a single project by its ID.
        
        **Authentication Required:** Yes (JWT)
        
        **Authorization:**
        - **ADMIN**: Can view any project
        - **USER**: Can only view projects they are assigned to
      operationId: getProjectById
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Project ID (UUID)
          schema:
            type: string
            format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
      responses:
        '200':
          description: Project retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
              examples:
                project:
                  summary: Project details
                  value:
                    id: 550e8400-e29b-41d4-a716-446655440000
                    name: Website Redesign
                    description: Complete redesign of the company website
                    createdById: 770e8400-e29b-41d4-a716-446655440002
                    createdAt: "2024-01-15T10:30:00Z"
                    updatedAt: "2024-01-15T10:30:00Z"
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '403':
          description: Forbidden - User not assigned to this project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenError'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Projects
      summary: Update a project
      description: |
        Update an existing project.
        
        **Authentication Required:** Yes (JWT)
        
        **Authorization:** ADMIN only
        
        - Only admins can update projects
        - All fields are optional (partial update)
      operationId: updateProject
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Project ID (UUID)
          schema:
            type: string
            format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
            examples:
              updateName:
                summary: Update project name
                value:
                  name: Website Redesign v2.0
              updateDescription:
                summary: Update description
                value:
                  description: Complete redesign with new branding guidelines
              updateAll:
                summary: Update multiple fields
                value:
                  name: Website & Mobile Redesign
                  description: Complete redesign across all platforms
                  userIds:
                    - 550e8400-e29b-41d4-a716-446655440000
      responses:
        '200':
          description: Project updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '403':
          description: Forbidden - User role not permitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenError'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Projects
      summary: Delete a project
      description: |
        Delete a project by ID.
        
        **Authentication Required:** Yes (JWT)
        
        **Authorization:** ADMIN only
        
        - Only admins can delete projects
        - Cascading delete: All associated tasks will be deleted
      operationId: deleteProject
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Project ID (UUID)
          schema:
            type: string
            format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
      responses:
        '200':
          description: Project deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Project deleted successfully
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '403':
          description: Forbidden - User role not permitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenError'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /projects/{id}/assign-users:
    post:
      tags:
        - Projects
      summary: Assign users to a project
      description: |
        Add users to a project team.
        
        **Authentication Required:** Yes (JWT)
        
        **Authorization:** ADMIN only
        
        - Only admins can assign users
        - Existing assignments are not duplicated
      operationId: assignUsersToProject
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Project ID (UUID)
          schema:
            type: string
            format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignUsersRequest'
            examples:
              assignMultiple:
                summary: Assign multiple users
                value:
                  userIds:
                    - 550e8400-e29b-41d4-a716-446655440000
                    - 660e8400-e29b-41d4-a716-446655440001
              assignSingle:
                summary: Assign single user
                value:
                  userIds:
                    - 550e8400-e29b-41d4-a716-446655440000
      responses:
        '200':
          description: Users assigned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '403':
          description: Forbidden - User role not permitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenError'
        '404':
          description: Project or user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /projects/{id}/remove-users:
    post:
      tags:
        - Projects
      summary: Remove users from a project
      description: |
        Remove users from a project team.
        
        **Authentication Required:** Yes (JWT)
        
        **Authorization:** ADMIN only
        
        - Only admins can remove users
        - Users are unassigned from the project
      operationId: removeUsersFromProject
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Project ID (UUID)
          schema:
            type: string
            format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignUsersRequest'
            examples:
              removeMultiple:
                summary: Remove multiple users
                value:
                  userIds:
                    - 550e8400-e29b-41d4-a716-446655440000
                    - 660e8400-e29b-41d4-a716-446655440001
      responses:
        '200':
          description: Users removed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
        '403':
          description: Forbidden - User role not permitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForbiddenError'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ========== TASK ENDPOINTS ==========
...